Create README.md
In this project I created a restful API for connected MongoDB database.If you want to make a sample project, you can use similar codes. If you want to run and try the project, the project will not run because I am working locally and there is no data in the database as I defined.